Post notify events to the torque_sockets event queue.
I am noticing that in a given pair of peers, A and B, if B
continuously spams A with torque_connection_send_to, but A does not
send any data beyond what the backend would normally send, eventually
A sends a notify packet for *every* packet it receives, and B does not
appear to *ever* send a notify packet.  It is not clear to me that
this is desired behavior.