torque_sockets client_identity is now the client_identity calculated by the server instead of the nonce.
The client generates its own nonce, which can easily result in nonce
collisions if the prngs of two different clients are in the same state.
The client_identity calculated by the server is guaranteed to be unique
for each connection, does not treat the client as a fully authoritative
source, and is a uint32 instead of being a uint64, which greatly simplifies
the torque_sockets API and allows us to use only builtin types.