Introduced connetions, take 1
I have opted to completely bypass the existing 'arranged connection'
as it's really rather silly to have a different handshake just because the
connection has a slightly different origin.

The new introduced connection protocol is a prefix to the standard
connection protocol.  The initiator sends a request to the introducer to
be introduced to a specific client identity.  The introducer then requests
that the initiator and the intended remote host send punch packets to each
other.  When the initiator receives the incoming punch packet it begins a
standard connection request to the originating address of the received
punch packet.

Concerns: At this point we do not ack or resend the introduction request,
the punch packet request, or the punch packets.  There are no security
measures in place, so if an interface were to receive an arbitrary punch
packet it would attempt to connect to the originating address.